Student results system
This application allows users to input students with their subject
results and determine the average for each subject (english, maths,
science) as well as the average mark across all tests for each student.
Along with this, other data such as class average information will also
be displayed in this application.